+++ Assets/Dust/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added DuActionMoveBy & DuActionMoveTo
 - Added DuActionDelay
 - Added DuActionCallback
+- Add onCompleteCallback for each DuAction
 
 ### Fixed